Update file(s): 2025-05-29

+++ exec/xmsdos.cpp
@@ -360,7 +360,7 @@ XBinary::_MEMORY_MAP XMSDOS::getMemoryMap(XBinary::MAPMODE mapMode, PDSTRUCT *pP
     qint64 nMaxOffset = (quint32)get_e_cp() * 0x200 - ((-get_e_cblp()) & 0x1ff);
 
     qint64 nHeaderOffset = 0;
-    qint64 nHeaderSize = (quint16)(get_e_cparhdr() * 16);
+    qint64 nHeaderSize = (quint32)(get_e_cparhdr() * 16);
     // qint64 nCodeOffset = (get_e_cparhdr() * 16) + (get_e_cs() * 16);
 
     // result.nEntryPointAddress = ((get_e_cs() << 16) + get_e_ip()) & 0xFFFFFFFF;
